requirements-dev.txt
-e ./
black
codespell
flake8
isort
mypy
pytest
pytest-cov
pytest-django
sphinx
sphinx-wagtail-theme
twine
wheel
